# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Sistema de Reservas - Universidad",
        "description": "Backend para la gestión de reservas de recursos (salas, equipos, etc.) con autenticación JWT y roles admin/usuario.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/notificaciones": {
            "get": {
                "tags": [
                    "Notificaciones"
                ],
                "summary": "Listar notificaciones del usuario",
                "description": "Obtiene la lista de notificaciones del usuario autenticado, ordenadas por fecha de creación (más recientes primero). Permite filtrar solo las no leídas.",
                "operationId": "4a5488228077528a278bea003da9a3cc",
                "parameters": [
                    {
                        "name": "solo_no_leidas",
                        "in": "query",
                        "description": "Filtrar solo notificaciones no leídas",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "example": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de notificaciones",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 5
                                            },
                                            "tipo": {
                                                "type": "string",
                                                "example": "reserva_creada"
                                            },
                                            "titulo": {
                                                "type": "string",
                                                "example": "Reserva creada correctamente"
                                            },
                                            "mensaje": {
                                                "type": "string",
                                                "example": "Has reservado el recurso Aula 101 del 2025-12-10 08:00:00 al 2025-12-10 10:00:00."
                                            },
                                            "leida": {
                                                "type": "boolean",
                                                "example": false
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-12-07T10:30:00.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/notificaciones/{id}/leer": {
            "put": {
                "tags": [
                    "Notificaciones"
                ],
                "summary": "Marcar notificación como leída",
                "description": "Marca una notificación específica del usuario autenticado como leída. Solo el propietario de la notificación puede realizar esta acción.",
                "operationId": "a18ce04f7d4e5dae08a80cb06665f37f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la notificación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notificación marcada como leída",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Notificación marcada como leída"
                                        },
                                        "notificacion": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "tipo": {
                                                    "type": "string",
                                                    "example": "reserva_creada"
                                                },
                                                "titulo": {
                                                    "type": "string",
                                                    "example": "Reserva creada correctamente"
                                                },
                                                "mensaje": {
                                                    "type": "string",
                                                    "example": "Has reservado el recurso Aula 101."
                                                },
                                                "leida": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-12-07T10:30:00.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Notificación no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/notificaciones/marcar-todas-leidas": {
            "put": {
                "tags": [
                    "Notificaciones"
                ],
                "summary": "Marcar todas las notificaciones como leídas",
                "description": "Marca todas las notificaciones no leídas del usuario autenticado como leídas en una sola operación.",
                "operationId": "69dbf08b85ef7395bb89708a2f1e8d46",
                "responses": {
                    "200": {
                        "description": "Todas las notificaciones marcadas como leídas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Todas las notificaciones han sido marcadas como leídas"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Introduce el token JWT con el formato: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}